requirements.txt
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
imageio==2.5.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
mock==3.0.4
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.7.1
pyparsing==2.4.0
PyQt5==5.12.1
PyQt5-sip==4.19.15
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
